feat: add error handling + getVideoByUsuario

+++ DB/controllers/videos.controller.js
@@ -11,9 +11,23 @@ const getVideos = async (_, res) => {
 
 const getVideo = async (req, res) => {
     const id = req.params.id;
+    if (!id) return res.status(400).json({ message: "An id is required" });
+
     try {
         const video = await videosService.getVideoById(id);
         return res.json(video);
+    } catch (error) {
+        return res.status(500).json({ error: error.message });
+    }
+};
+
+const getVideoByUsuario = async (req, res) => {
+    const id_usuario = req.userId;
+    if (!user) return res.status(400).json({ message: "A user is required" });
+
+    try {
+        const video = await videosService.getVideoByUsuario(id_usuario);
+        return res.json(video)
     } catch (error) {
         return res.status(500).json({ error: error.message });
     }
@@ -25,6 +39,8 @@ const uploadVideo = async (req, res) => {
 
 const deleteVideo = async (req, res) => {
     const id = req.params.id;
+    if (!id) return res.status(400).json({ message: "An id is required" });
+
     try {
         const deletedId = await videosService.deleteVideo(id);
         return res.json(deletedId);
